requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
discord.py==1.2.5
idna==2.8
multidict==4.7.2
python-dotenv==0.10.3
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
Wand==0.5.8
websockets==6.0
yarl==1.4.2
